#jira none
Some warning/error messages about using very long paths include the full path (which is a very long path) in the message itself, making the message too long to be displayed when used e.g., in dialog windows (e.g., in File --> Save Current As). Updated the messages from something like "Can't convert the path %s because it is too long" into something like "Can't convert the path because it is too long (%d characters). [...] Full path: %s". This preserves the file path value, but only at the end of the message, so the UI dialogs can display everything (but the file path itself) and at the same time console warnings can display the whole information.

#rb none

[CL 10095760 by Gines Hidalgo in Dev-Editor branch]