#jira UE-80859
Bug fixed: Loading an Editor Layout used to close all open sub-editors.  Now, a new Dialog warning window is displayed if any sub-editor instance or PIE/SIE are opened when the user tries to load a new layout, asking whether he want to continue loading the layout (that would lead to automatically close+- those without saving them).

Note: Previously, all the load options were disabled while PIE was running. However, this could be confusing for the user, who might not know why they are disabled and not associate it with PIE being running (Although it was shown in the Load Layout tooltip, but people is not likely to read those...). Instead, this warning is merged with the one for the opened asset editors.

This is a quick fix is for 4.24, to avoid deep code changes. However, for later versions, the following tasks should be done as well:
- Investigate whether PIE or the sub-editor can be maintained opened, rather than closing them when a new layout is loaded.

#rb chris.gagnon

[CL 10108258 by Gines Hidalgo in Dev-Editor branch]