Adds a new default initialization mode to Niagara module inputs where you can bind to available compatible parameters via a dropdown menu in the Selected Details panel inside the graph editor.
By selecting a Module parameter in the Parameters window there's now a new DefaultMode widget that can be "Value", "Binding" or "Custom".

When "Value" is selected it will display an appropriate widget below it to set the default value.
When "Binding" is selected it will display a dropdown widget of a list of available names to bind. Only built-in names and names explicitly added to the graph or Parameters list will show up
When "Custom is selected it will display neither of the widgets, indicating that the intiialization is to be done via a sub-graph. This will be streamlined later when the default pins are removed in favor of a separate initialization node.

The robustness of how all of this interacts with rapid iteration parameters requires a bit more work, but this is a good checkpoint as any. With respect to UE-82847, the bindings behave the same as any other non-value overrides for now.

Includes an accompanying autotest to test different permutations of initializing in the graph and overriding in the stack.

#rb shaun.kime, nicholas.goldstein


#ROBOMERGE-SOURCE: CL 10111573 via CL 10111577
#ROBOMERGE-BOT: (v587-10111126)

[CL 10111583 by morten vassvik in Main branch]