#jira UE-81469
New feature: Added the capability of adding Radio buttons to the submenu main button (not only to its entries).

This CL is the first of 2 CLs that introduce multiple improvements on the "View Mode" menu of the Viewport, including a new feature request from JIRA UE-81469. However, we isolate this particular new feature into an isolated CL given that it can be applicable anywhere in the Editor, not only for "View Mode".

Issue: If any subcategory of "View Mode" was chosen, the user could not know in which category it was located (unless the user manually looked into each category until he found the selected one).
Solution: Added the checked functionality (with a radio button) to each one of the categories, behaving similarly to any subcategory. This way, both the radio button of the category and the one from the specific subcategory will be checked simulatenously.

More technical details:
Adding this functionality to "View Mode" means adding the functionality of "FUIAction::FIsActionChecked" into "AddSubMenu", which includes modifications to "MultiBox", "SMenuEntryBlock", and "MultiBoxBuilder".

1) We are changing a submenu so it has a radioButton
2) To specify a radio button a developer has to use a different API that includes a delegate (FUIAction) to say if it should be checked or not before this CL
3) The different API for AddSubMenu was not passing along the command list for that child submenu to use
4) The updated code now passes the command list along for the child submenu to use

#rb rex.hill

[CL 10228324 by Gines Hidalgo in Dev-Editor branch]