Removed utility method for determining EngineModeEx.
On Linux/Mac EngineModeEx was not properly set. This was because FGenericPlatformCrashContext::Initialize was called during PreInit before all the information needed had been setup. EngineMode and EngineModeEx are now updated by a callback. WindowsPlatformCrashContext was calling Initialize to support crashes during static initialization, which was refactored to only call it when actually needed.

#jira UE-82522, UE-82525
#rb brandon.schaefer

#ROBOMERGE-SOURCE: CL 10284648 in //UE4/Release-4.24/...
#ROBOMERGE-BOT: RELEASE (Release-4.24 -> Main) (v591-10236483)

[CL 10284649 by johan berg in Main branch]