#jira UE-81469
Multiple improvements to the "View Mode" menu of the Viewport, including a feature request from JIRA UE-81469. Most of these improvements are related to the submenu of "View Mode" called "Buffer Visualization":

Issue 1/6 (the JIRA itself): The "View Mode" UI displays as its title the name on the selected category or subcategory chosen by the user (or otherwise the default one). However, this behavior was not consistent across different categories. For the "Buffer Visualization" category, if any of its subcategories were chosen, "View Mode" was only displaying "Buffer Visualization" as title, rather than the particular subcategory chosen. This was not the case for the other 2 categories ("Level of Detail" and "Optimization Viewmode").
Solution 1: The displayed name of "View Mode" is now forced to be the exact chosen (sub)category, regardless of the category ("Buffer Visualization", "Level of Detail", or "Optimization Viewmode").

Issue 2/6: The name displayed as the selected "View Mode" was being assigned in parallel with the displayed name of the buttons representing each (sub)category. This resulted in some categories and subcategories having different names in the buttons than in the "View Mode" label itself. The behavior was also not consistent across the different menus inside "View Mode".
Solution 2: The displayed name of the chosen label is forced to match the displayed name of the button itself.

Issue 3/6: "Optimization Viewmodes" was the only category without an icon.
Solution 3: Icon assigned to one of the icons of its subcategories. In particular, the one from "Quad Overdraw".

Issue 4/6: "Buffer visualization" was the only category without a proper tooltip, it was displaying a FName, rather than a more human-readable FText.
Solution 4: Assigned its tooltip to the FText field of DisplayedName.

Issue 5/6: When the subcategory "Overview" (of "Buffer Visualization") was chosen, the displayed name on each individual subwindow that are generated corresponded with limited FNames (rathern than a more human-readable text).
Solution 5: Assigned their value to the FTexts corresponding to their DisplayedName value.

Issue 6/6: Inside the possible Buffer Visualization Materials, `LightingModel` and `ShadingModel` are being used interchangeably all over the code, leading to missmatches and confusion for the user. E.g., `UMaterialInterface` and `UMaterial` use `LightingModel`, while `FBufferVisualizationData` (thus `FBufferVisualizationMenuCommands` and `FEditorViewportClient`) use `ShadingModel`.
Solution 6: Assigned all of the above classes to the same value.

Issue and solution 7 added in previous CL # 10228324.

#rb chris.gagnon

[CL 10286305 by Gines Hidalgo in Dev-Editor branch]