Removed utility method for determining EngineModeEx.
On Linux/Mac EngineModeEx was not properly set. This was because FGenericPlatformCrashContext::Initialize was called during PreInit before all the information needed had been setup. EngineMode and EngineModeEx are now updated by a callback. WindowsPlatformCrashContext was calling Initialize to support crashes during static initialization, which was refactored to only call it when actually needed.

#jira UE-82522, UE-82525
#rb brandon.schaefer

#ROBOMERGE-OWNER: ryan.gerleve
#ROBOMERGE-AUTHOR: johan.berg
#ROBOMERGE-SOURCE: CL 10284648 in //UE4/Release-4.24/... via CL 10284649
#ROBOMERGE-BOT: NETWORKING (Main -> Dev-Networking) (v593-10286020)

[CL 10294154 by johan berg in Dev-Networking branch]