Fixed: Enabling Fixed Frame Rate doesn't Lock the Frame Rate
This was a regression introduced in CL 9842913 while trying to fix negative time deltas when switching from a fixed time step to variable (UE-81145). However, using bTimeWasManipulated to detect the switch is incorrect, because that's always true when fixed frame rate is enabled. The code now stores the previous state of bUseFixedTimeStep explicitly and compares against that to determine when the interval needs to be reset.

#jira UE-84752
#lockdown cristina.riveron
#rb Matt.Hoffman
#rnx

#ROBOMERGE-SOURCE: CL 10341305 in //UE4/Release-4.24/...
#ROBOMERGE-BOT: RELEASE (Release-4.24 -> Main) (v594-10333955)

[CL 10341323 by mihnea balta in Main branch]