Adds a new default initialization mode to Niagara variables where you can bind to available compatible parameters via a dropdown menu in the Selected Details panel inside the graph editor.
By selecting a variable in the Parameters window there's now a new DefaultMode widget that can be "Value", "Binding" or "Custom".

When "Value" is selected it will display an appropriate widget below it to set the default value. A value can only be defined if that variable is referenced in the graph.
When "Binding" is selected it will display a dropdown widget of a list of available names to bind. Only built-in names and names explicitly added to the graph or Parameters list will show up
When "Custom is selected it will display neither of the widgets, indicating that the intiialization is to be done via a sub-graph. This will be streamlined later when the default pins are removed in favor of a separate initialization node.

The bindings should be functionally equivalent to the old version of setting the default with a sub-graph and the Begin Defaults node.

Includes two autotest to test different permutations of initializing in the graph and overriding in the stack.

Passes preflights. All autotests pass locally.

#rb shaun.kime, nicholas.goldstein

#ROBOMERGE-OWNER: robert.manuszewski
#ROBOMERGE-AUTHOR: morten.vassvik
#ROBOMERGE-SOURCE: CL 10590589 via CL 10590606 via CL 10590609
#ROBOMERGE-BOT: CORE (Main -> Dev-Core) (v610-10636431)

[CL 10694465 by morten vassvik in Dev-Core branch]