Persistent IDs for GPU particles.
"Requires persistent IDs" is now supported for GPU emitters. Each emitter maintains a list of free IDs and an ID to index mapping table. The ID to index mapping is filled in by the simulation step. The free IDs are computed after each simulation step by compacting the slots in the ID mapping table which are not touched by the simulation.

#rb stu.mckenna, shaun.kime

#ROBOMERGE-OWNER: ryan.gerleve
#ROBOMERGE-AUTHOR: mihnea.balta
#ROBOMERGE-SOURCE: CL 10808036 via CL 10808037 via CL 10808038
#ROBOMERGE-BOT: NETWORKING (Main -> Dev-Networking) (v610-10636431)

[CL 10839860 by mihnea balta in Dev-Networking branch]