Implemented the particle read data interface on GPU.
The method UNiagaraDataInterface::GetFunctionHLSL receives more information about the function being generated, such as the specifiers and the index inside the list of functions produced by the DI. The extra information is also stored inside FNiagaraDataInterfaceGPUParamInfo, so it's accessible in the Bind() method of the DI's compute shader. This allows the particle read DI to retrieve the field names for each function when the shader is bound, and to map them to register indices when the shader is used. The register indices are cached, so that they're only computed when the source emitter changes.

Also moved the GPU free ID list out of FNiagaraDataBuffer and into FNiagaraDataSet. The previous code was wrong, because when the particle count changes, we need to resize the *input* free ID list, not only the output. Having a list per data buffer doesn't make sense, and a single list per emitter is easier to manage. This fixes duplicated IDs at the beginning of the simulation, and other problems which appeared whenever the buffer had to grow.

#rb shaun.kime, stu.mckenna, simon.tovey


#ROBOMERGE-OWNER: mihnea.balta
#ROBOMERGE-AUTHOR: mihnea.balta
#ROBOMERGE-SOURCE: CL 10975676 via CL 10975687
#ROBOMERGE-BOT: (v632-10940481)

[CL 10975760 by mihnea balta in Main branch]