Added methods to extend the layout of Unreal Insights
Added configuration structures for major & minor tabs. Currently minor tabs are only customized in the timing view.
Removed built-in behaviour for session connection. In editor connection is now initialized in the same manner as in the standalone tool (from the GameplayInsights plugin).
Fixed issue with trace services where the analysis service would not be created.

#rb Ionut.Matasaru

[CL 11028750 by Thomas Sarkanen in Dev-Core branch]