Optimized array memory. Modified DefaultCalculateSlackGrow to not add slack for the first allocation if NumElements exceeds FirstGrow. Various places were calling SetNum() and never growing beyond that, which meant the slack was never used.
Saved 70-90MB in Fortnite on consoles.

I gave this a lot of local testing due to reports of memory corruption with a similar change (10207395), but I didn't reproduce any issues locally, either in replays or the automated tests. It's been tested on Win64 and all consoles. This change is different to 10207395 in that it's simpler and applies to all containers, so it's possible the issues arose from a bug in the original change.

#rb steve.robb


#ROBOMERGE-SOURCE: CL 11065948 via CL 11065980 via CL 11065992
#ROBOMERGE-BOT: (v637-11041722)

[CL 11066003 by ben woodhouse in Main branch]