Fix FName number parsing from unterminated strings
Copy the number to a temporary buffer until we have an overload of Atoi64 that takes the input length. The generic implementation of Atoi64 for TCHAR converts the input to ASCII then converts that to an integer. For unterminated strings, this could cause an access violation or convert an obscenely large input to ASCII.

#rb Johan.Torp, Matt.Peters


#ROBOMERGE-SOURCE: CL 11201262 via CL 11201265
#ROBOMERGE-BOT: (v640-11091645)

[CL 11201273 by devin doucette in Main branch]