Fix asset view with an active filter killing the editor performance when assets are added.
In a project with a lot of assets (let's say ~400K assets), SAssetView::ProcessRecentlyAddedAssets() would process a single asset during it's alloted time slice. The recreation of the set on each call would make the editor run at >10fps until the whole array is processed, at one asset per tick.

It's easy to reproduce by starting the editor and clicking on the "Content" root in the Content Browser, then typing anything in the Filter box while the asset discovery is in progress. The asset discovery will complete but the editor performance will be degraded until the filter is cleared.

With these changes, all new assets are processed in the allowed time with a single call to ProcessRecentlyAddedAssets()
#rb jamie.dale


#ROBOMERGE-SOURCE: CL 11292517 via CL 11292526 via CL 11292528 via CL 11292529
#ROBOMERGE-BOT: (v647-11244347)

[CL 11293016 by sebastien lussier in Main branch]