Fixed: if multiple instances of a DI were used in a script, the HLSL translator produced all function combinations for each instance.
DataInterfaceRegisteredFunctions accumulated all registered functions from all the instances, then the generator iterated over it for each instance, which was incorrect. The information was already available per instance in FNiagaraScriptDataInterfaceCompileInfo.RegisteredFunctions, so we can just use that and remove DataInterfaceRegisteredFunctions entirely.

Already submitted this once as 11386264, but that introduced a new bug, because RegisteredFunctions can contain duplicates. Did it properly this time, so that GetFunctionHLSL is called once per unique function instance.

#rb nicholas.goldstein


#ROBOMERGE-SOURCE: CL 11416526 via CL 11416570
#ROBOMERGE-BOT: (v654-11333218)

[CL 11416580 by mihnea balta in Main branch]