Heavily optimize level streaming (Chaos AABBTree) in Reverb (~15x faster - Forgotten City sublevel make visible ~5min -> ~20sec).
The Chaos AABBTree implementation has a heuristic that reoptimizes the tree every X elements, where X defaults to 10k. The reoptimize tree pass is very expensive, as it does a ton of heap allocations and copies (due to resizing). In just the Forgotten City sub level group, ReoptimizeTree() is called ~118 times, as there are 1.2mil elements. Since the size of FElement is 64bytes, the heuristic is now set to 128MiB of pending data size, which handles 2mil elements before any reoptimize occurs.

In addition to the heuristic, this change also fixes a high number of heap allocations and copies from resize. For example, the original code, per leaf, would do a heap allocation in GatherElements() for MGlobalPayloads.Num(), but then it would reserve that num + MDirtyElements.Num(), which would result in a new larger allocation and then a copy of the old contents into the new allocation. Since OutElements is passed in by reference, I also added a global reserve sum that I compute and size that entire array for the entire collection of leaves, as well as a few other similar optimizations.

#rb jaco.vandyk
#[fyi] brian.karis, michal.valient, michael.lentine, devon.penney


#ROBOMERGE-SOURCE: CL 11414176 via CL 11419533 via CL 11419545
#ROBOMERGE-BOT: (v654-11333218)

[CL 11419594 by jaco vandyk in Main branch]