Fix unexpected state in ULevelStreaming::UpdateStreamingState
The FLevelStreamingWrapper::operator< relies on the streaming level priority for its comparison. The issue is that TMap using this class as a key relied on that operator too for the initial insertion, and since the priority of items already in the map could change afterward, it would break the ordering and prevent proper indexing (for example, Find, Contains, Removed could fail to find items that were actually in the map).

We can actually simply use a map of ULevelStreaming* for the LevelsToProcess map. The insertion in the priority sorted StreamingLevels array is performed using Algo::LowerBound() anyway.

#rb marc.audy, jeanfrancois.dube


#ROBOMERGE-SOURCE: CL 11547315 via CL 11547321 via CL 11547325
#ROBOMERGE-BOT: (v654-11333218)

[CL 11547330 by sebastien lussier in Main branch]