Merge 11407298 to Release-4.25.
Fix geometry collection transforms.  Fix a bug with collision particles including the origin.  Fix a problem with the geometry collection componet not being delivered simulation results, and having to compute transforms itself.  Fix an issue where it seems transforms and bounds were computed on data for the prior frame rather that the current.  Implement support for "None" as an option for geometry collection collision geometry.  Add a cvar for disabling gravity for geometry collections.  Code cleanup and compaction.  Do array preallocations where we can.

#fyi: Devon.Penney, Michael.Lentine, Brice.Criswell
#rb: none
#jira: UESP-2072

[CL 11555980 by Ryan Kautzman in 4.25 branch]