Merge from partner-magicleap-staging: Fixes in AugmentedReality impl for MagicLeap.
Reset worker thread setup for image tracking so it restarts on consequtive VRPreviews.
We need to UpdateTrackedGeometry in FLuminARTrackedPointResource::UpdateGeometryData even if we couldn't get a valid position and orientation to update to.  This can sometimes happen after a level transition, and if we don't UpdateTrackedGeometry, the Point may not have a valid ARSystem pointer, which may crash if certain blueprint functions (GetLocalToWorldTransform, for example) are called at this time.

Removed GetTrackableHandleMap from the ARTrackingSystem interface. Instead, a new function RemoveHandleIf is used to set the 'StoppedTracking' state and remove the handle.
RemoveHandleIf takes in a higher-order function that is applied to every stored handle if the geometry type matches. When the applied function returns true, the tracked handle stops tracking and is removed internally. Users of the tracking system can perform additional removal behavior as they see fit via the input function.
#jira UE-89351
#rb steve.smith

[CL 11579592 by steve smith in 4.25 branch]