Fix issue where new/changed screenshots could not be approved in the editor on some platforms (Mac, Xbox, others?)
The issue was that the code that captures screenshots on a device and sends them to the editor was providing a path that represented what a screenshot *would* be saved at on the local system. (The screenshot was not actually saved, it's returned as raw data in the message).

This meant that the editor would try to convert this to a relative path for its needs which would fail as the the path roots were differenti (e.g. g:\settings\ on xbox).

The fixes here are -

1) Add some warnings to the code that tries to save screenshots if that fails

2) Change the names of the screenshot functions in AutomationCommon to clarify what they should be used for.

3) Change the 'path' variable in the message to screenshot name.

4) Calculate the path to use for screenshots on the editor.


#tests ran EngineTest on WIn64, PS4, and XboxOne (which now passes).
#jira UE-69399
#rb swarm

[CL 11733277 by Andrew Grant in 4.25 branch]