Fix crash when undoing curve edits
Switched curve viewer to use 'named curve' model the same as the curve editor. This is needed because the tree view doesnt refresh straight away, so tracks dont get removed until the tick after an undo is fired, while the undo event gets pushed to the track even when it isnt hooked up to the tree. Using the named curve approach means that the invalid track doesnt end up accessing a deleted curve.

#jira UE-88277 - Crash in Editor After Editing an Animation in the Curve Editor and Rapidly Pressing CTRL + Z
#rb Jurre.deBaare

#ROBOMERGE-SOURCE: CL 11820722 in //UE4/Release-4.25/... via CL 11820774
#ROBOMERGE-BOT: RELEASE (Release-4.25Plus -> Main) (v656-11643781)

[CL 11820778 by thomas sarkanen in Main branch]