Integrating the dataprep preview system.
The way the dataprep objects notify the systems that they were edited was change for the need of the preview system. The parameterization code was updated.
The asset preview widget was refactor to support columns with a custom sorting. This also put it in a better to support custom filter at some point.
Functions of the dataprep filters were modified to process on an array view instead of a full array. Also added some functions to output the result and some additional info for the preview system to some buffer instead of returning an array.
Updated the dataprep editor so that our graph widgets can have a optional weak reference to the dataprep editor. (While at it I also fixed alpha issue of one of the images we are using for our new nodes)

This doesn't include the ui and the code to clear a preview.

#jira UEENT-3333
#jira UEENT-3334
#rb JeanLuc.Corenthin

[CL 11873215 by Julien StJean in 4.25 branch]