Fixed crash when recompiling Niagara systems
FNiagaraShaderMapId objects were not hashed correctly because of a typo introduced with the loading optimization changes, so GIdToNiagaraShaderMap was broken. Using GetTypeHash() on FSHAHash::Hash is wrong, because that's just a uint8[20], so the function call resolves to the void* overload and it returns different values for the same hash. FSHAHash has its own GetTypeHash() overload, so we need to use that.

Also discovered that BaseCompileHash was always all-zeroes because GPU scripts don't have a graph representation. Made UNiagaraGraph::RebuildCachedCompileIds() use the spawn script hash in that case, so we have something to use as a hash for the keys in GIdToNiagaraShaderMap.

#jira UE-89503
#rb Shaun.Kime

#ROBOMERGE-OWNER: ryan.gerleve
#ROBOMERGE-AUTHOR: mihnea.balta
#ROBOMERGE-SOURCE: CL 12002668 in //UE4/Release-4.25/... via CL 12002685 via CL 12002721
#ROBOMERGE-BOT: NETWORKING (Main -> Dev-Networking) (v656-11643781)

[CL 12039099 by mihnea balta in Dev-Networking branch]