Consolidate GeometryCollection storage of implicit shapes into a single ManagedArray of TSharedPtr<FImplicitObject> - non thread safe variant.  Removes SharedGeometryAttribute and SharedImplicitsAttribute.  This likely means that legacy geometry collections will need to be re-generated after this poin.
Copy implicit shapes so the physics thread has it's own copy of the data, for the time being for clarity.  Add a DeepCopy() function to TImplicitObject, where the base implementation simply forwards to Copy(), which is implemented for all implicit shapes except TLevelSet.  Implement TLevelSet::DeepCopy(), leaving TLevelSet::Copy() unimplemented, which asserts in the base implementation.

Add a cvar in GeometryCollectionPhysicsProxy that overrides collision filtering, and enables collisions for geometry collections against everything.  This cvar IS CURRENTLY ENABLED BY DEFAULT so that tests pass while proper support for collision filters is implemented for geometry collections.

These changes get all currently active geometry collection simulation tests passing, and geometry collections simulating in the editor again, although, not colliding.

(All of this was talked over with Brice.)

[FYI] Michael.Lentine, Benn.Gallagher. Brice.Criswell, Bill.Henderson
#rb: none
#jira: none

#ROBOMERGE-SOURCE: CL 12047778 in //UE4/Release-4.25/...
#ROBOMERGE-BOT: RELEASE (Release-4.25 -> Release-4.25Plus) (v656-11643781)

[CL 12047784 by ryan kautzman in 4.25-Plus branch]